fix: `MsgCreateValidator` and `MsgUpdateValidator` race condition (#225)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Description
This PR fixes the `MsgCreateValidator` and `MsgUpdateValidator` race condition bug by changing how we handle those types of messages. Instead of storing the infos directly from them, we use gRPC to query the entire data for the height that those msgs were transmitted to, and then store such information. This way even when a `MsgUpdateValidator` is sent, the entire data will be fetched and stored avoiding partial data storage that caused the bug. 

Fix #219 

## Checklist
- [x] Targeted PR against correct branch.
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work.
- [ ] Wrote unit tests.  
- [x] Re-reviewed `Files changed` in the Github PR explorer.